Changed the signature to manage template
